Extract region fetch pipeline
